+++ chromium/content/browser/cache_storage/cache_storage_dispatcher_host.cc
@@ -4,8 +4,12 @@
 
 #include "content/browser/cache_storage/cache_storage_dispatcher_host.h"
 
+#include <stddef.h>
+#include <utility>
+
 #include "base/bind.h"
 #include "base/logging.h"
+#include "base/macros.h"
 #include "base/strings/string16.h"
 #include "base/strings/utf_string_conversions.h"
 #include "base/trace_event/trace_event.h"
@@ -15,14 +19,15 @@
 #include "content/browser/cache_storage/cache_storage_manager.h"
 #include "content/common/cache_storage/cache_storage_messages.h"
 #include "content/public/browser/content_browser_client.h"
+#include "content/public/common/origin_util.h"
 #include "storage/browser/blob/blob_data_handle.h"
-#include "third_party/WebKit/public/platform/WebServiceWorkerCacheError.h"
+#include "third_party/WebKit/public/platform/modules/serviceworker/WebServiceWorkerCacheError.h"
 
 namespace content {
 
 namespace {
 
-const uint32 kFilteredMessageClasses[] = {CacheStorageMsgStart};
+const uint32_t kFilteredMessageClasses[] = {CacheStorageMsgStart};
 
 blink::WebServiceWorkerCacheError ToWebServiceWorkerCacheError(
     CacheStorageError err) {
@@ -41,6 +46,10 @@ blink::WebServiceWorkerCacheError ToWebServiceWorkerCacheError(
   }
   NOTREACHED();
   return blink::WebServiceWorkerCacheErrorNotImplemented;
+}
+
+bool OriginCanAccessCacheStorage(const GURL& url) {
+  return IsOriginSecure(url);
 }
 
 }  // namespace
@@ -104,6 +113,10 @@ void CacheStorageDispatcherHost::OnCacheStorageHas(
     const GURL& origin,
     const base::string16& cache_name) {
   TRACE_EVENT0("CacheStorage", "CacheStorageDispatcherHost::OnCacheStorageHas");
+  if (!OriginCanAccessCacheStorage(origin)) {
+    bad_message::ReceivedBadMessage(this, bad_message::CSDH_INVALID_ORIGIN);
+    return;
+  }
   context_->cache_manager()->HasCache(
       origin, base::UTF16ToUTF8(cache_name),
       base::Bind(&CacheStorageDispatcherHost::OnCacheStorageHasCallback, this,
@@ -117,6 +130,10 @@ void CacheStorageDispatcherHost::OnCacheStorageOpen(
     const base::string16& cache_name) {
   TRACE_EVENT0("CacheStorage",
                "CacheStorageDispatcherHost::OnCacheStorageOpen");
+  if (!OriginCanAccessCacheStorage(origin)) {
+    bad_message::ReceivedBadMessage(this, bad_message::CSDH_INVALID_ORIGIN);
+    return;
+  }
   context_->cache_manager()->OpenCache(
       origin, base::UTF16ToUTF8(cache_name),
       base::Bind(&CacheStorageDispatcherHost::OnCacheStorageOpenCallback, this,
@@ -130,6 +147,10 @@ void CacheStorageDispatcherHost::OnCacheStorageDelete(
     const base::string16& cache_name) {
   TRACE_EVENT0("CacheStorage",
                "CacheStorageDispatcherHost::OnCacheStorageDelete");
+  if (!OriginCanAccessCacheStorage(origin)) {
+    bad_message::ReceivedBadMessage(this, bad_message::CSDH_INVALID_ORIGIN);
+    return;
+  }
   context_->cache_manager()->DeleteCache(
       origin, base::UTF16ToUTF8(cache_name),
       base::Bind(&CacheStorageDispatcherHost::OnCacheStorageDeleteCallback,
@@ -141,6 +162,10 @@ void CacheStorageDispatcherHost::OnCacheStorageKeys(int thread_id,
                                                     const GURL& origin) {
   TRACE_EVENT0("CacheStorage",
                "CacheStorageDispatcherHost::OnCacheStorageKeys");
+  if (!OriginCanAccessCacheStorage(origin)) {
+    bad_message::ReceivedBadMessage(this, bad_message::CSDH_INVALID_ORIGIN);
+    return;
+  }
   context_->cache_manager()->EnumerateCaches(
       origin,
       base::Bind(&CacheStorageDispatcherHost::OnCacheStorageKeysCallback, this,
@@ -155,7 +180,10 @@ void CacheStorageDispatcherHost::OnCacheStorageMatch(
     const CacheStorageCacheQueryParams& match_params) {
   TRACE_EVENT0("CacheStorage",
                "CacheStorageDispatcherHost::OnCacheStorageMatch");
-
+  if (!OriginCanAccessCacheStorage(origin)) {
+    bad_message::ReceivedBadMessage(this, bad_message::CSDH_INVALID_ORIGIN);
+    return;
+  }
   scoped_ptr<ServiceWorkerFetchRequest> scoped_request(
       new ServiceWorkerFetchRequest(request.url, request.method,
                                     request.headers, request.referrer,
@@ -163,13 +191,14 @@ void CacheStorageDispatcherHost::OnCacheStorageMatch(
 
   if (match_params.cache_name.empty()) {
     context_->cache_manager()->MatchAllCaches(
-        origin, scoped_request.Pass(),
+        origin, std::move(scoped_request),
         base::Bind(&CacheStorageDispatcherHost::OnCacheStorageMatchCallback,
                    this, thread_id, request_id));
     return;
   }
   context_->cache_manager()->MatchCache(
-      origin, base::UTF16ToUTF8(match_params.cache_name), scoped_request.Pass(),
+      origin, base::UTF16ToUTF8(match_params.cache_name),
+      std::move(scoped_request),
       base::Bind(&CacheStorageDispatcherHost::OnCacheStorageMatchCallback, this,
                  thread_id, request_id));
 }
@@ -192,7 +221,7 @@ void CacheStorageDispatcherHost::OnCacheMatch(
       new ServiceWorkerFetchRequest(request.url, request.method,
                                     request.headers, request.referrer,
                                     request.is_reload));
-  cache->Match(scoped_request.Pass(),
+  cache->Match(std::move(scoped_request),
                base::Bind(&CacheStorageDispatcherHost::OnCacheMatchCallback,
                           this, thread_id, request_id, cache));
 }
@@ -203,9 +232,29 @@ void CacheStorageDispatcherHost::OnCacheMatchAll(
     int cache_id,
     const ServiceWorkerFetchRequest& request,
     const CacheStorageCacheQueryParams& match_params) {
-  // TODO(gavinp,jkarlin): Implement this method.
-  Send(new CacheStorageMsg_CacheMatchAllError(
-      thread_id, request_id, blink::WebServiceWorkerCacheErrorNotImplemented));
+  IDToCacheMap::iterator it = id_to_cache_map_.find(cache_id);
+  if (it == id_to_cache_map_.end()) {
+    Send(new CacheStorageMsg_CacheMatchError(
+        thread_id, request_id, blink::WebServiceWorkerCacheErrorNotFound));
+    return;
+  }
+
+  scoped_refptr<CacheStorageCache> cache = it->second;
+  if (request.url.is_empty()) {
+    cache->MatchAll(
+        base::Bind(&CacheStorageDispatcherHost::OnCacheMatchAllCallback, this,
+                   thread_id, request_id, cache));
+    return;
+  }
+
+  scoped_ptr<ServiceWorkerFetchRequest> scoped_request(
+      new ServiceWorkerFetchRequest(request.url, request.method,
+                                    request.headers, request.referrer,
+                                    request.is_reload));
+  cache->Match(
+      std::move(scoped_request),
+      base::Bind(&CacheStorageDispatcherHost::OnCacheMatchAllCallbackAdapter,
+                 this, thread_id, request_id, cache));
 }
 
 void CacheStorageDispatcherHost::OnCacheKeys(
@@ -330,7 +379,7 @@ void CacheStorageDispatcherHost::OnCacheStorageMatchCallback(
   }
 
   if (blob_data_handle)
-    StoreBlobDataHandle(blob_data_handle.Pass());
+    StoreBlobDataHandle(*blob_data_handle);
 
   Send(new CacheStorageMsg_CacheStorageMatchSuccess(thread_id, request_id,
                                                     *response));
@@ -350,9 +399,50 @@ void CacheStorageDispatcherHost::OnCacheMatchCallback(
   }
 
   if (blob_data_handle)
-    StoreBlobDataHandle(blob_data_handle.Pass());
+    StoreBlobDataHandle(*blob_data_handle);
 
   Send(new CacheStorageMsg_CacheMatchSuccess(thread_id, request_id, *response));
+}
+
+void CacheStorageDispatcherHost::OnCacheMatchAllCallbackAdapter(
+    int thread_id,
+    int request_id,
+    const scoped_refptr<CacheStorageCache>& cache,
+    CacheStorageError error,
+    scoped_ptr<ServiceWorkerResponse> response,
+    scoped_ptr<storage::BlobDataHandle> blob_data_handle) {
+  scoped_ptr<CacheStorageCache::Responses> responses(
+      new CacheStorageCache::Responses);
+  scoped_ptr<CacheStorageCache::BlobDataHandles> blob_data_handles(
+      new CacheStorageCache::BlobDataHandles);
+  if (error == CACHE_STORAGE_OK) {
+    DCHECK(response);
+    responses->push_back(*response);
+    if (blob_data_handle)
+      blob_data_handles->push_back(*blob_data_handle);
+  }
+  OnCacheMatchAllCallback(thread_id, request_id, cache, error,
+                          std::move(responses), std::move(blob_data_handles));
+}
+
+void CacheStorageDispatcherHost::OnCacheMatchAllCallback(
+    int thread_id,
+    int request_id,
+    const scoped_refptr<CacheStorageCache>& cache,
+    CacheStorageError error,
+    scoped_ptr<CacheStorageCache::Responses> responses,
+    scoped_ptr<CacheStorageCache::BlobDataHandles> blob_data_handles) {
+  if (error != CACHE_STORAGE_OK && error != CACHE_STORAGE_ERROR_NOT_FOUND) {
+    Send(new CacheStorageMsg_CacheMatchAllError(
+        thread_id, request_id, ToWebServiceWorkerCacheError(error)));
+    return;
+  }
+
+  for (const storage::BlobDataHandle& handle : *blob_data_handles)
+    StoreBlobDataHandle(handle);
+
+  Send(new CacheStorageMsg_CacheMatchAllSuccess(thread_id, request_id,
+                                                *responses));
 }
 
 void CacheStorageDispatcherHost::OnCacheKeysCallback(
@@ -406,15 +496,14 @@ void CacheStorageDispatcherHost::DropCacheReference(CacheID cache_id) {
 }
 
 void CacheStorageDispatcherHost::StoreBlobDataHandle(
-    scoped_ptr<storage::BlobDataHandle> blob_data_handle) {
-  DCHECK(blob_data_handle);
+    const storage::BlobDataHandle& blob_data_handle) {
   std::pair<UUIDToBlobDataHandleList::iterator, bool> rv =
       blob_handle_store_.insert(std::make_pair(
-          blob_data_handle->uuid(), std::list<storage::BlobDataHandle>()));
-  rv.first->second.push_front(storage::BlobDataHandle(*blob_data_handle));
+          blob_data_handle.uuid(), std::list<storage::BlobDataHandle>()));
+  rv.first->second.push_front(storage::BlobDataHandle(blob_data_handle));
 }
 
-void CacheStorageDispatcherHost::DropBlobDataHandle(std::string uuid) {
+void CacheStorageDispatcherHost::DropBlobDataHandle(const std::string& uuid) {
   UUIDToBlobDataHandleList::iterator it = blob_handle_store_.find(uuid);
   if (it == blob_handle_store_.end())
     return;
